// Hawa-front/src/app/shared/model/master-data/master-data-all.model.ts
export class MasterDataAll {
    actorTypes: [
      {
        id: 0;
        name: string;
        code: string;
        acronymName: string
      }
    ];
    contactTypes: [
      {
        key: string;
        code: string;
        text: string
      }
    ];
    landUseCerts: [
      {
        key: string;
        code: string;
        text: string
      }
    ];
    userTypes: [
      {
        key: string;
        code: string;
        text: string
      }
    ];
    userStatues: [
      {
        key: string;
        code: string;
        text: string
      }
    ];
    forestCerts: [
      {
        key: string;
        code: string;
        text: string
      }
    ];
    forestplotReliability: [
      {
        key: string;
        code: string;
        text: string
      }
    ];
    compartments: [
      {
        key: string;
        code: string;
        text: string
      }
    ];
    actorRoles: [
      {
        key: string;
        code: string;
        text: string
      }
    ];
    contactStatuses: [
      {
        key: string;
        code: string;
        text: string
      }
    ]
  }